+++ src/mmz/mod.rs
@@ -97,14 +97,45 @@ pub enum DelimKind {
     Both,
 }
 
-/// Operator precedence for user-declared notation.
+/// Operator precedence for user-declared notation. 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub enum Prec {
     Num(u32),
     Max,
 }
 
-/// 
+/// The corresponding term number, its ident (name), whether it's right-associative, and any
+/// literal items/symbols used in the declaration. For general notation declarations that 
+/// start with a non-variable, that gets left out of `lits`, and instead becomes
+/// one of the keys in `prefixes`. When the math parser finds the initial token, it
+/// looksup the term using the `prefixes` mapping, and then uses the remaining consts
+/// as checks to make sure the item is syntactically correct, but it will have already
+/// parsed and gone past the initial const token.
+///
+/// Examples of lits:
+///```text
+/// NotationInfo { 
+///     term_num: 2, 
+///     term_ident: an, 
+///     rassoc: false, 
+///     lits: [Var { pos: 0, prec: Num(34) }, Const(/\), Var { pos: 1, prec: Num(35) }] 
+/// }
+///
+/// NotationInfo { 
+///    term_num: 1, 
+///    term_ident: not, 
+///    rassoc: true, 
+///    lits: [Var { pos: 0, prec: Num(41) }] 
+///}
+///
+/// // notice the missing `Const([)` at the front; it's a key in the `prefixes` map.
+/// NotationInfo { 
+///     term_num: 14, 
+///     term_ident: sb, 
+///     rassoc: true, 
+///     lits: [Var { pos: 0, prec: Num(1) }, Const(/), Var { pos: 1, prec: Num(1) }, Const(]), Var { pos: 2, prec: Num(42) }] 
+/// }
+///```
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct NotationInfo<'a> {
     pub term_num: u32,
@@ -118,6 +149,12 @@ pub enum Fix {
     Infixl,
     Infixr,
     Prefix,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Assoc {
+    Left,
+    Right
 }
 
 /// Inside of a general notation declaration, you can either have variables (`Var`) or notation symbols.
@@ -155,6 +192,20 @@ pub enum Coe {
     },
 }
 
+/// The persistent (remains between the checking of each mm0 declaration) mm0 state.
+/// 
+/// The most complex part of this is the notation system. Users can declare 4 kinds of notation,
+/// prefix, infixl, infixr, and general. The verifier tracks these in only two maps, putting 
+/// prefix and general notations in the `prefixes` map, and infixl/infixr declarations in the `infixes` map.
+/// The reason for this grouping is because the math parser looks for these two things separately, and the places
+/// where a prefix item might be used are also the places where a general notation can be, and the same is true
+/// for both flavors of infix.
+///
+/// There are four maps holding notation information. 
+/// 1. `consts`: map of the item's name to its associated precedence.
+/// 2. `occupied_precs`: Prevents parser ambiguities; see the doc comment above the field for more info about this.
+/// 3. `prefixes`: Associates the items' name to its `NotationInfo { term_num, term_ident, rassoc, lits }`
+/// 3. `infixes`: Associates the items' name to its `NotationInfo { term_num, term_ident, rassoc, lits }`
 pub struct MmzMem<'a> {
     pub outline: &'a Outline<'a>,
 
@@ -164,12 +215,65 @@ pub struct MmzMem<'a> {
     pub mmz_pos: usize,
     pub delims: Vec<(u8, DelimKind)>,
 
-    /// Maps the string representing a user-declared notation to its associated precedence value.
-    /// `/\` |-> `Prec::Num(34)`
+    /// Maps the string representing a user-declared notation to its associated precedence value. For example:
+    ///```text 
+    /// peano.mm0/and 
+    ///
+    /// /\ |-> `Prec::Num(34)`
+    ///```
     pub consts: HashMap<Str<'a>, Prec>,
-    /// A map containing prefix notations
+    /// A map containing both prefix and general notation declarations. 
+    ///
+    /// `Ident |-> NotationInfo { term_num, term_ident, rassoc, lits }`
     pub prefixes: HashMap<Str<'a>, NotationInfo<'a>>,
     pub infixes: HashMap<Str<'a>, NotationInfo<'a>>,    
+    /// A mapping of precedence levels to (prefix \/ infixl \/ infixr).
+    /// This is necessary to prevent ambiguities in the math parser; while the math parser 
+    /// won't completely break without this, "You will get _a_ parse, but part of the idea here 
+    /// is that you will also get _the_ parse", where _the_ parse is the one that conforms to
+    /// the context-free grammar laid out [here](https://github.com/digama0/mm0/blob/master/mm0.md#secondary-parsing).
+    ///
+    /// If the environment permits both an infixl and infixr at the same precednce level, for example:
+    ///
+    ///```text
+    /// infixr im: $->$ prec 25;
+    /// infixl an: $/\$ prec 25;
+    ///```
+    ///
+    /// Then beginning from state `Start`, and using rules 0, 1, and 2 taken from the mm0 expression grammar, we can derive 
+    /// both `(a -> (b /\ c))` and `((a -> b) /\ c)` as parses of `$ a -> b /\ c $`:
+    ///```text
+    /// Start := $ expr(0) $;
+    /// rule0 := expression(p1) -> expression(p2)                   (if p1 < p2)
+    /// rule1 := expression(p) -> expression(p) OP expression(p+1)  (if OP is infixl prec p)
+    /// rule2 := expression(p) -> expression(p+1) OP expression(p)  (if OP is infixr prec p)
+    ///
+    /// with both:
+    /// infixr im: $->$ prec 25;
+    /// infixl an: $/\$ prec 25;
+    ///
+    /// To get ((a -> b) /\ c):
+    /// ($ a -> b /\ c $, 0)   ~rule0~> ($ a -> b /\ c $, 25) 
+    /// ($ a -> b /\ c $, 25)  ~rule1~> ($ a -> b $, 25) /\ ($ c $, 26)
+    /// ($ a -> b $, 25)       ~rule2~> ($ a $, 26) -> ($ b $, 25)
+    ///
+    /// To get (a -> (b /\ c)):
+    /// ($ a -> b /\ c $, 0)  ~rule0~> ($ a -> b /\ c $, 25) by rule0
+    /// ($ a -> b /\ c $, 25) ~rule2~> ($ a $, 26) -> ($ b /\ c $, 25) by rule2
+    /// ($ b /\ c $, 25)      ~rule1~> ($ b $, 25) /\ ($ c $, 26) by rule1
+    ///```
+    ///
+    /// By disallowing both `-> infixr` and `/\ infixl` to inhabit the same precedence, this ambiguity
+    /// can be avoided.
+    ///
+    /// `peano.mm1` and the mm1 files based on it adopt a convention of using even precedence numbers for 
+    /// left-associative infix operators, and odd precedence numbers for right-associative infix operators
+    /// as a simple way to avoid this kind of collision.
+    /// infixl is L
+    /// infixr is R
+    /// (prefix w/ arity > 0) /\ general ending with a variable are Rassoc
+    /// 0ary prefix /\ general ending with a const don't affect the map.
+    pub occupied_precs: HashMap<Prec, Assoc>,    
     // unneeded
     //pub declared_notations: FxIndexMap<Str<'a>, DeclaredNotation<'a>>,
     pub nonlocal_termdefs: HashMap<TermIdent<'a>, TermNum>,
@@ -183,6 +287,7 @@ pub struct MmzMem<'a> {
     /// A map of sorts `s` to some sort `t` such that `t` is provable and `c: s -> t` is in `coes`,
     /// if one exists.
     pub coe_prov: HashMap<SortNum, SortNum>,
+
     num_sorts: u8,
     num_termdefs: u32,
     num_asserts: u32,
@@ -206,6 +311,7 @@ impl<'a> MmzMem<'a> {
             consts: HashMap::new(),
             prefixes: HashMap::new(),
             infixes: HashMap::new(),
+            occupied_precs: HashMap::new(),
             //declared_notations: FxIndexMap::with_hasher(Default::default()),
             nonlocal_termdefs: HashMap::new(),
             nonlocal_asserts: HashMap::new(),
@@ -299,7 +405,15 @@ impl<'a> MmzMem<'a> {
         let term_num = none_err!(self.nonlocal_termdefs.get(ident).copied())?;
         self.outline.get_term_by_num(term_num)
     }        
-  
+
+    pub fn occupy_prec(&mut self, term_ident: Str<'a>, prec: Prec, assoc: Assoc) -> Res<()> {
+        if let Some(already) = self.occupied_precs.insert(prec, assoc) {
+            if already != assoc {
+                return Err(VerifErr::Msg(format!("{:?}; Cannot re-occupy prec {:?} with different associativity", term_ident, prec)))
+            }
+        }
+        Ok(())
+    }
 }
 
 //#[derive(Debug)]